refactor: basic table layout
